feat: add redux slice for repos
The app now employs useGithub hook to get the repos, without any filtering or sorting
and then we dispatch an action to put all of these repos into the redux state.
Then, we use local component state and the sortRepos hook to sort
the repos in the way we want. We also use the filter to
do similar work.